+++ cli/doctor.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	docker "github.com/fsouza/go-dockerclient"
-	"github.com/robfig/cron/v3"
+	"github.com/netresearch/go-cron"
 
 	"github.com/netresearch/ofelia/core"
 )